Taking advantage of the RNG
The chained shiny mode repeated every result. That was due to the fact
that this RNG (and every 32-bit LCG) returns the same numbers but
switching the most significant digit if the RNG state is the same but
its most significant digit is not (that is a consequence of the fact
that the nth least significant bit has a period of 2^n or less).

This was sorted out by taking advantage of that fact. Now, only half of
the possible seeds are considered, and every seed not lower than 0x8000
0000 is checked only if itself minus 0x8000 0000 matches the wanted IVs
and the mode is not the IVs to chained shiny PID one. As a result,
everything that gets a PID from IVs is twice as fast.